Don't stop demod when issuing DiSEqC or Tone Burst commands
This may prevent locking depending on which order are called tuning
and DiSEqC commands.

Here are comments from Ralph (Digital Devices):

The macro stops the demod, because a sudden loss of signal, like
you get when switching inputs with diseqc command, can cause the
TS output section in the stv0900 to lose packets on the other demod.
But this does not happen very often (see errrata of stv0900).
I guess it depends on in what order you use diseqc and tuning that it
could cause trouble.